evaluation

+++ src/deepsysid/networks/switching.py
@@ -201,11 +201,11 @@ class UnconstrainedSwitchingLSTM(SwitchingBaseLSTM):
 
         states = torch.zeros(
             size=(batch_size, sequence_length, self.state_dim), device=control.device
-        )
+        ).float()
         if previous_state is None:
             state = torch.zeros(
                 size=(batch_size, self.state_dim), device=control.device
-            )
+            ).float()
             state[:, : self.output_dim] = previous_output
         else:
             state = previous_state
@@ -340,11 +340,11 @@ class StableSwitchingLSTM(SwitchingBaseLSTM):
 
         states = torch.zeros(
             size=(batch_size, sequence_length, self.state_dim), device=control.device
-        )
+        ).float()
         if previous_state is None:
             state = torch.zeros(
                 size=(batch_size, self.state_dim), device=control.device
-            )
+            ).float()
             state[:, : self.output_dim] = previous_output
         else:
             state = previous_state
